# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenTogg API",
        "description": "Time tracking REST API. Authenticate with a Bearer token created in Settings.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/export": {
            "get": {
                "tags": [
                    "Data Portability"
                ],
                "summary": "Export all user data",
                "description": "Returns all vectors, tags, and time entries as JSON for backup or migration.",
                "operationId": "b3354398c5619b834abe2d042b6daf73",
                "responses": {
                    "200": {
                        "description": "Full data export",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "version": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "exported_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "vectors": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "time_entries": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/import": {
            "post": {
                "tags": [
                    "Data Portability"
                ],
                "summary": "Import user data",
                "description": "Import vectors, tags, and time entries. Uses external_id for upsert — safe for syncing between instances.",
                "operationId": "07a4385b09486b898c987f1919dcb4ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "version"
                                ],
                                "properties": {
                                    "version": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vectors": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "external_id",
                                                "name"
                                            ],
                                            "properties": {
                                                "external_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "color": {
                                                    "type": "string",
                                                    "example": "#6B7280"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "external_id",
                                                "name"
                                            ],
                                            "properties": {
                                                "external_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "name": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "time_entries": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "external_id",
                                                "started_at"
                                            ],
                                            "properties": {
                                                "external_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "vector_external_id": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "tag_external_ids": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "started_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "stopped_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import complete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "imported": {
                                            "properties": {
                                                "vectors": {
                                                    "type": "integer"
                                                },
                                                "tags": {
                                                    "type": "integer"
                                                },
                                                "time_entries": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}